refactor: 파일 이름 변경 및 함수, 상수  분리
hash.js -> path.js

fetchServer.js 생성
