+++ src/ts/constant/path.js
@@ -1,3 +1,10 @@
+export const BASE_SERVER_URL = 'http://localhost:8080';
+
+export const SERVER_PATH = {
+  REGISTER: 'register',
+  LOGIN: 'login',
+};
+
 export const HASH = {
   ITEM_MANAGE: '#item-manage',
   COIN_RECHARGE: '#coin-recharge',
